# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Db server API description.",
        "title": "Db server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://openitstudio.ru",
            "email": "fedor@support-pc.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/admin/ds": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "List data sources",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data source",
                    "Admin"
                ],
                "summary": "List data sources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.DataSource"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Create data source",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data source",
                    "Admin"
                ],
                "summary": "Create data source",
                "parameters": [
                    {
                        "description": "Data source info",
                        "name": "ds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.DataSource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.DataSource"
                        }
                    }
                }
            }
        },
        "/admin/ds/dse/{dsId}": {
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Create data source endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data source",
                    "Admin"
                ],
                "summary": "Create data source endpoint",
                "parameters": [
                    {
                        "description": "Data source info",
                        "name": "dse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.DataSourceEndpoint"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Data source id",
                        "name": "dsId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.DataSourceEndpoint"
                        }
                    }
                }
            }
        },
        "/admin/ds/dse/{dsId}/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Data source endpoint detail info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data source",
                    "Admin"
                ],
                "summary": "Data source endpoint info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Data source id",
                        "name": "dsId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Endpoint id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.DataSource"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Update date source endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data source",
                    "Admin"
                ],
                "summary": "Update date source endpoint",
                "parameters": [
                    {
                        "description": "Endpoint info",
                        "name": "dse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.DataSourceEndpoint"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Endpoint id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Data source id",
                        "name": "dsId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.DataSourceEndpoint"
                        }
                    }
                }
            }
        },
        "/admin/ds/{dsIid}/dse": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "List data source endpoints",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data source",
                    "Admin"
                ],
                "summary": "List data source endpoints",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Data source id",
                        "name": "dsIid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.DataSourceEndpoint"
                            }
                        }
                    }
                }
            }
        },
        "/admin/ds/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Data source detail info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data source",
                    "Admin"
                ],
                "summary": "Data source info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.DataSource"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Update date source",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data source",
                    "Admin"
                ],
                "summary": "Update date source",
                "parameters": [
                    {
                        "description": "Source info",
                        "name": "source",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.DataSource"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Source info",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.DataSource"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Delete data source",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data source",
                    "Admin"
                ],
                "summary": "Delete data source",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ds id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
...
        "/ds/dse/{dsId}/{id}": {
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Delete data source endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data source",
                    "Admin"
                ],
                "summary": "Delete data source endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Data source endpoint id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Data source id",
                        "name": "dsId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
        "/dse/{id}": {
            "get": {
                "description": "Get data source by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data source",
                    "Public Api"
                ],
                "summary": "Get item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth key",
                        "name": "db-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Source id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/project.Project"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.DataSource": {
            "type": "object",
            "properties": {
                "cache": {
                    "description": "Cache result in local db",
                    "type": "boolean"
                },
                "dsn": {
                    "description": "Data source dsn",
                    "type": "string"
                },
                "endpoints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.DataSourceEndpoint"
                    }
                },
                "id": {
                    "description": "The data source UUID\nexample: 6204011c-30e6-408b-8aaa-dd8219860b4b",
                    "type": "string"
                },
                "project": {
                    "description": "Linked project",
                    "$ref": "#/definitions/project.Project"
                },
                "project_id": {
                    "description": "Linked project  UUID\nexample: 6204011c-30e6-408b-8aaa-dd8214860b4b",
                    "type": "string"
                },
                "title": {
                    "description": "Data source title",
                    "type": "string"
                },
                "type": {
                    "description": "The data source type\nEnum of DsType\nexample: Mysql",
                    "type": "string"
                }
            }
        },
        "ds.DataSourceEndpoint": {
            "type": "object",
            "properties": {
                "dataSource": {
                    "description": "Linked data source",
                    "$ref": "#/definitions/ds.DataSource"
                },
                "data_source": {
                    "description": "Linked data source UUID\nexample: 6204011c-33e6-408b-8aaa-dd8214860b4b",
                    "type": "string"
                },
                "id": {
                    "description": "The data source endpoint UUID\nexample: 6234011c-30e6-408b-8aaa-dd8219860b4b",
                    "type": "string"
                },
                "table_name": {
                    "description": "Endpoint table name",
                    "type": "string"
                },
                "title": {
                    "description": "Data source endpoint title",
                    "type": "string"
                }
            }
        },
...
        "project.Project": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "origins": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}